@@ -20,10 +20,11 @@
 	"devDependencies": {
 		"@ianvs/prettier-plugin-sort-imports": "^4.3.1",
 		"@playwright/test": "^1.43.0",
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/adapter-node": "^5.0.1",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/adapter-cloudflare": "^4.7.4",
+		"@sveltejs/adapter-node": "^5.2.9",
+		"@sveltejs/kit": "^2.8.1",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@tailwindcss/nesting": "^0.0.0-insiders.565cd3e",
 		"@types/eslint": "^8.56.0",
 		"@types/markdown-it": "^14.0.0",
@@ -40,7 +41,7 @@
 		"electron-builder": "^25.0.3",
 		"eslint": "^8.56.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
+		"eslint-plugin-svelte": "^2.45.1",
 		"highlight.js": "^11.9.0",
 		"lucide-svelte": "^0.372.0",
 		"markdown-it": "^14.1.0",
@@ -51,19 +52,19 @@
 		"postcss-load-config": "^5.0.2",
 		"postcss-nested": "^6.2.0",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"prettier-plugin-tailwindcss": "^0.5.9",
 		"semver": "^7.6.3",
-		"svelte": "^4.2.19",
-		"svelte-check": "^3.6.0",
+		"svelte": "^5.2.2",
+		"svelte-check": "^4.0.0",
 		"svelte-sonner": "^0.3.27",
 		"tailwind-scrollbar": "^3.1.0",
 		"tailwindcss": "^3.3.6",
 		"thememirror": "^2.0.1",
 		"tslib": "^2.4.1",
 		"typesafe-i18n": "^5.26.2",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.3",
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4",
 		"vite-node": "^2.0.5"
 	},
 	"optionalDependencies": {
