@@ -1,6 +1,7 @@
 import type { BaseTranslation } from '../i18n-types';
 
 const en = {
+	addConnection: 'Add connection',
 	allowConnections: 'Change your server settings to allow connections from',
 	allowMixedContent: 'Allow mixed content',
 	apiKey: 'API key',
@@ -21,9 +22,12 @@ const en = {
 	checkNow: 'Check now',
 	clear: 'Clear',
 	confirmDeletion: 'Confirm deletion',
-	connect: 'Connect',
-	connected: 'Connected',
+	connectionType: 'Connection type',
+	connectionLabelHelp: 'Identifies the connection in the models list',
 	connecting: 'Connecting...',
+	connectionIsVerified: 'Connection has been verified and is ready to use',
+	connectionFailedToVerify:
+		'Connection failed to verify, check the connection settings and try again',
 	content: 'Content',
 	controls: 'Controls',
 	controlsOnlyAvailableForOllama:
@@ -39,6 +43,7 @@ const en = {
 	deleteAllSessions: 'Delete all sessions',
 	deleteAllSettings: 'Delete all settings',
 	deleteKnowledge: 'Delete knowledge',
+	deleteServer: 'Delete server',
 	deleteSession: 'Delete session',
 	disconnected: 'Disconnected',
 	dismiss: 'Dismiss',
@@ -61,6 +66,7 @@ const en = {
 	isLatestVersion: 'A newer version is available',
 	knowledge: 'Knowledge',
 	knowledgeSaved: 'Knowledge saved',
+	label: 'Label',
 	language: 'Language',
 	lastUsedModels: 'Recently used models',
 	light: 'Light',
@@ -75,6 +81,8 @@ const en = {
 	modelNotFound: 'Model not found',
 	modelOptions: 'Model options',
 	modelWasDownloaded: '{model:string} was downloaded',
+	modelsFilter: 'Model names filter',
+	modelsFilterHelp: "Hides model names that don't start with the filter",
 	motd: 'Motd',
 	name: 'Name',
 	newKnowledge: 'New knowledge',
@@ -82,6 +90,7 @@ const en = {
 	noApiKey: 'No API key',
 	noKnowledgeChosen: 'Create new knowledge or choose one from the list',
 	noRecentModels: 'No recently used models',
+	noServerConnections: 'No server connections, add one to start',
 	noSessionChosen: 'Create a new session or choose an existing one from the list',
 	notFound: 'The page you are looking for does not exist',
 	numBatch: 'Num batch',
@@ -91,9 +100,10 @@ const en = {
 	numPredict: 'Num predict',
 	numThread: 'Number of threads',
 	numa: 'NUMA',
+	ollama: 'Ollama',
 	ollamaLibrary: "Ollama's library",
-	openaiSyncFailed: 'Failed to connect to OpenAI. Check your API key or network connection',
-	openaiSyncSuccessful: 'Sync was successful. OpenAI models are now available in Sessions',
+	openAIOfficialAPI: 'OpenAI: Official API',
+	openAICompatible: 'OpenAI: Compatible servers (i.e. llama.cpp)',
 	otherModels: 'Other models',
 	penalizeNewline: 'Penalize newline',
 	presencePenalty: 'Presence penalty',
@@ -109,13 +119,16 @@ const en = {
 	repeatPenalty: 'Repeat penalty',
 	retry: 'Retry',
 	run: 'Run',
+	reVerify: 'Re-verify',
 	runtimeOptions: 'Runtime options',
 	save: 'Save',
 	search: 'Search',
 	searchEmpty: 'No results found',
 	seed: 'Seed',
 	seeDocs: 'See docs',
-	server: 'Server',
+	servers: 'Servers',
+	serverSettingsUpdated:
+		'Server configuration was updated and needs to be re-verified in "Settings"',
 	session: 'Session',
 	sessions: 'Sessions',
 	settings: 'Settings',
@@ -135,6 +148,8 @@ const en = {
 	typicalP: 'Typical P',
 	useMlock: 'Use MLOCK',
 	useMmap: 'Use MMAP',
+	useModelsFromThisServer: 'Use models from this server',
+	verify: 'Verify',
 	version: 'Version',
 	vocabOnly: 'Vocab only',
 	writePromptToStart: 'Write a prompt to start a new session',
