@@ -87,7 +87,8 @@ const ja = {
 	search: '検索',
 	searchEmpty: '結果が見つかりませんでした',
 	seeDocs: 'ドキュメントを参照する',
-	server: 'サーバー',
+	servers: 'サーバー',
+	serverSettingsUpdated: 'サーバーの設定が更新され、"設定"で再確認が必要です',
 	session: 'セッション',
 	sessions: 'セッション',
 	settings: '設定',
