@@ -3,8 +3,10 @@ import type { BaseTranslation, Translation } from '../i18n-types';
 
 const zh_cn = {
 	...(en as Translation),
+	addConnection: '添加连接',
 	allowConnections: '更改您的服务器设置以允许连接：',
 	allowMixedContent: '允许混合内容',
+	apiKey: 'API 密钥',
 	areYouSureYouWantToDeleteAllKnowledge: '确定删除所有的知识吗？',
 	areYouSureYouWantToDeleteAllSessions: '确定删除所有的会话吗？',
 	areYouSureYouWantToDeleteAllSettings: '确定删除所有的设置吗？',
@@ -13,6 +15,7 @@ const zh_cn = {
 	automatic: '自动',
 	automaticallyCheckForUpdates: '自动检查更新',
 	availableModels: '可用模型',
+	baseUrl: '基础 URL',
 	browseModels: '浏览可用模型',
 	cancel: '取消',
 	cantConnectToOllamaServer: '无法连接到 Ollama 服务器',
@@ -21,9 +24,14 @@ const zh_cn = {
 	checkNow: '立刻检查',
 	clear: '清除',
 	confirmDeletion: '确定删除',
-	connected: '已连接',
+	connectionType: '连接类型',
+	connectionLabelHelp: '在模型列表中标识连接',
+	connecting: '连接中...',
+	connectionIsVerified: '连接已验证并可以使用',
+	connectionFailedToVerify: '连接验证失败，请检查连接设置后重试',
 	content: '内容',
 	controls: '控制器',
+	controlsOnlyAvailableForOllama: '高级控制目前仅适用于 Ollama 模型',
 	copy: '复制',
 	couldntCheckForUpdates: '无法自动检查更新',
 	couldntConnectToOllamaServer: '无法连接到 Ollama 服务器',
@@ -35,6 +43,7 @@ const zh_cn = {
 	deleteAllSessions: '删除所有会话',
 	deleteAllSettings: '删除所有设置',
 	deleteKnowledge: '删除知识',
+	deleteServer: '删除服务器',
 	deleteSession: '删除会话',
 	disconnected: '已断开连接',
 	dismiss: '忽略',
@@ -43,11 +52,11 @@ const zh_cn = {
 	emptyKnowledge: '无知识',
 	emptySessions: '无会话',
 	error: '错误',
-	frequencyPenalty: 'Frequency penalty',
 	genericError: '抱歉，发生了错误',
 	goToDownloads: '前往下载页',
 	goToKnowledge: '前往知识页',
 	goToReleases: '前往发布页',
+	howToObtainOpenAIKey: '如何获取 OpenAI API 密钥？',
 	howToUpdateDocker: '如何更新 Docker 容器？',
 	interface: '界面',
 	internalServerError: '发生了内部错误，请稍后重试',
@@ -55,51 +64,63 @@ const zh_cn = {
 	isLatestVersion: '有新版本可用',
 	knowledge: '知识',
 	knowledgeSaved: '知识已保存',
+	label: '标签',
 	language: '语言',
 	lastUsedModels: '最近使用的模型',
 	light: '浅色',
 	messageOfTheDay: '资讯',
 	messages: '消息',
+	modelNotFound: '未找到模型',
 	modelOptions: '模型设置',
 	modelWasDownloaded: '{model:string} 已下载',
+	modelsFilter: '模型名称过滤',
+	modelsFilterHelp: '隐藏不以过滤器开头的模型名称',
 	motd: '资讯',
 	name: '名称',
 	newKnowledge: '新建知识',
 	newSession: '新建会话',
+	noApiKey: '无 API 密钥',
 	noKnowledgeChosen: '新建知识或从列表中选择',
 	noRecentModels: '无近期使用的模型',
+	noServerConnections: '无服务器连接，添加一个以开始',
 	noSessionChosen: '新建会话或从列表中选择已存在的会话',
 	notFound: '您正在查看的页面不存在',
+	ollama: 'Ollama',
 	ollamaLibrary: 'Ollama 库',
+	openAICompatible: 'OpenAI：兼容服务器（如 llama.cpp）',
+	openAIOfficialAPI: 'OpenAI：官方 API',
 	otherModels: '其他模型',
 	prompt: '提示词',
 	promptPlaceholder: '请输入任何内容',
 	pullModel: '下拉模型',
-	pullModelPlaceholder: '模型标签(例如 llama3.1)',
+	pullModelPlaceholder: '模型标签（例如 llama3.1）',
 	pullingModel: '正在下拉模型',
 	random: '随机',
 	refreshToUpdate: '刷新以更新',
 	releaseHistory: '发布历史',
-	repeatLastN: 'Repeat last N',
-	repeatPenalty: 'Repeat penalty',
 	retry: '重试',
+	reVerify: '重新验证',
 	run: '运行',
 	runtimeOptions: '运行时设置',
 	save: '保存',
 	search: '搜索',
 	searchEmpty: '无结果',
 	seeDocs: '查看文档',
-	server: '服务器',
+	servers: '服务器',
+	serverSettingsUpdated: '服务器配置已更新，需要在"设置"中重新验证',
 	session: '会话',
 	sessions: '会话',
 	settings: '设置',
 	somethingWentWrong: '抱歉，发生了错误',
+	stopCompletion: '停止完成',
 	stopResponse: '停止回复',
 	success: '成功',
 	system: '系统',
 	systemPrompt: '系统提示词',
 	tryingToConnectNotLocalhost:
 		'如果您想连接到 {hostname:string} 或 {ip:string} 上不可用的Ollama服务器，请尝试',
+	useModelsFromThisServer: '使用此服务器的模型',
+	verify: '验证',
 	version: '版本',
 	writePromptToStart: '写一段提示词以开始新会话',
 	you: '你'
