@@ -1,12 +1,13 @@
 import { expect, test } from '@playwright/test';
 
 import {
+	chooseFromCombobox,
 	MOCK_OPENAI_COMPLETION_RESPONSE_1,
 	MOCK_OPENAI_MODELS,
 	MOCK_SESSION_1_RESPONSE_1,
+	mockOllamaModelsResponse,
 	mockOpenAICompletionResponse,
-	mockOpenAIModelsResponse,
-	mockTagsResponse
+	mockOpenAIModelsResponse
 } from './utils';
 
 test.describe('OpenAI Integration', () => {
@@ -15,42 +16,47 @@ test.describe('OpenAI Integration', () => {
 	});
 
 	test('fails to fetch data with an incorrect API key', async ({ page }) => {
+		await chooseFromCombobox(page, 'Connection type', 'OpenAI: Official API');
+		await page.getByText('Add connection').click();
+		await expect(page.getByLabel('Base URL')).toHaveValue('https://api.openai.com/v1');
+
 		await page.route('**/v1/models', (route) => {
 			route.fulfill({ status: 401, json: { error: { message: 'Invalid API key' } } });
 		});
-
-		await page.getByLabel('Base URL').fill('https://api.openai.com/v1');
 		await page.getByLabel('API Key').fill('sk-invalidapikey');
-		await page.getByRole('button', { name: 'Connect' }).click();
+		await page.getByRole('button', { name: 'Verify' }).click();
 
-		await expect(page.getByText('Failed to connect to OpenAI')).toBeVisible();
+		await expect(
+			page.getByText('Connection failed to verify, check the connection settings and try again')
+		).toBeVisible();
 	});
 
 	test('handles network connection error', async ({ page }) => {
 		await page.route('**/v1/models', (route) => route.abort('failed'));
+		await chooseFromCombobox(page, 'Connection type', 'OpenAI: Official API');
+		await page.getByText('Add connection').click();
+		await expect(page.getByLabel('Base URL')).toHaveValue('https://api.openai.com/v1');
 
-		await page.getByLabel('Base URL').fill('https://api.openai.com/v1');
 		await page.getByLabel('API Key').fill('sk-validapikey');
-		await page.getByRole('button', { name: 'Connect' }).click();
-
-		await expect(page.getByText('Failed to connect to OpenAI')).toBeVisible();
+		await page.getByRole('button', { name: 'Verify' }).click();
+		await expect(
+			page.getByText('Connection failed to verify, check the connection settings and try again')
+		).toBeVisible();
 	});
 
-	test('cannot send fetch requests without apiKey or baseUrl set', async ({ page }) => {
-		const connectButton = page.getByRole('button', { name: 'Connect' });
-
-		await expect(connectButton).toBeDisabled();
-
-		await page.getByLabel('Base URL').fill('https://api.openai.com/v1');
-		await expect(connectButton).toBeDisabled();
+	test('cannot send fetch requests without a baseUrl set', async ({ page }) => {
+		await chooseFromCombobox(page, 'Connection type', 'OpenAI: Official API');
+		await page.getByText('Add connection').click();
+		const verifyButton = page.getByRole('button', { name: 'Verify' });
+		await expect(verifyButton).toBeEnabled();
+		await expect(page.getByLabel('Base URL')).toHaveValue('https://api.openai.com/v1');
 
 		await page.getByLabel('Base URL').clear();
-		await page.getByLabel('API Key').fill('sk-validapikey');
-		await expect(connectButton).toBeDisabled();
+		await expect(verifyButton).toBeDisabled();
 	});
 
 	test('models list is sorted correctly', async ({ page }) => {
-		await mockTagsResponse(page);
+		await mockOllamaModelsResponse(page);
 		await mockOpenAIModelsResponse(page, MOCK_OPENAI_MODELS);
 
 		await page.getByText('Sessions', { exact: true }).click();
@@ -90,7 +96,6 @@ test.describe('OpenAI Integration', () => {
 		await page.getByTestId('new-session').click();
 		await page.getByLabel('Available models').click();
 		await page.getByRole('option', { name: 'gpt-3.5-turbo' }).click();
-
 		await mockOpenAICompletionResponse(page, MOCK_OPENAI_COMPLETION_RESPONSE_1);
 
 		// Simulate sending a message
@@ -100,18 +105,18 @@ test.describe('OpenAI Integration', () => {
 
 		// Check localStorage
 		const sessions = await page.evaluate(() => localStorage.getItem('hollama-sessions'));
-		expect(sessions).toContain('"model":"gpt-3.5-turbo"');
+		expect(sessions).toContain('"name":"gpt-3.5-turbo"');
 	});
 
 	test('only GPT models are available in FieldModelSelect', async ({ page }) => {
 		await mockOpenAIModelsResponse(page, MOCK_OPENAI_MODELS);
 		expect(MOCK_OPENAI_MODELS).toHaveLength(3);
 		expect(MOCK_OPENAI_MODELS[2].id).toContain('text-davinci-003');
+		await expect(page.getByLabel('Model names filter')).toHaveValue('gpt');
 
 		await page.getByText('Sessions', { exact: true }).click();
 		await page.getByTestId('new-session').click();
 		await page.getByLabel('Available models').click();
-
 		await expect(page.getByRole('option', { name: 'gpt-3.5-turbo' })).toBeVisible();
 		await expect(page.getByRole('option', { name: 'gpt-4' })).toBeVisible();
 		await expect(page.getByRole('option', { name: 'text-davinci-003' })).not.toBeVisible();
