@@ -8,7 +8,7 @@ import {
 	MOCK_SESSION_1_RESPONSE_3,
 	MOCK_SESSION_2_RESPONSE_1,
 	mockCompletionResponse,
-	mockTagsResponse,
+	mockOllamaModelsResponse,
 	textEditorLocator
 } from './utils';
 
@@ -16,7 +16,7 @@ test.describe('Session', () => {
 	let promptTextarea: Locator;
 
 	test.beforeEach(async ({ page }) => {
-		await mockTagsResponse(page);
+		await mockOllamaModelsResponse(page);
 		promptTextarea = page.locator('.prompt-editor__textarea');
 	});
 
@@ -149,13 +149,9 @@ test.describe('Session', () => {
 		const settingsLink = page.locator('.layout__a', { hasText: 'Settings' });
 
 		await page.goto('/');
-		await expect(settingsLink).toHaveClass(/ layout__a--active/);
-		await expect(sessionLink).not.toHaveClass(/ layout__a--active/);
-
-		await sessionLink.click();
-		await expect(page.getByText('No sessions')).toBeVisible();
-		await expect(settingsLink).not.toHaveClass(/ layout__a--active/);
 		await expect(sessionLink).toHaveClass(/ layout__a--active/);
+		await expect(settingsLink).not.toHaveClass(/ layout__a--active/);
+		await expect(page.getByText('No sessions')).toBeVisible();
 		await expect(
 			page.locator('aside', {
 				hasText: 'Who would win in a fight between Emma Watson and Jessica Alba?'
@@ -718,12 +714,8 @@ test.describe('Session', () => {
 	});
 
 	test('handles errors when fetching models', async ({ page }) => {
-		await page.goto('/');
 		await page.getByText('Sessions', { exact: true }).click();
 		await page.getByTestId('new-session').click();
-
-		await expect(page.getByTestId('disconnected-server')).not.toBeVisible();
-
 		await chooseModel(page, MOCK_API_TAGS_RESPONSE.models[0].name);
 		await promptTextarea.fill('Who would win in a fight between Emma Watson and Jessica Alba?');
 		await expect(page.getByText('Run')).toBeEnabled();
@@ -735,8 +727,6 @@ test.describe('Session', () => {
 		await page.route('**/chat', async (route) => {
 			await route.abort('failed');
 		});
-		await page.getByTestId('new-session').click();
-		await chooseModel(page, MOCK_API_TAGS_RESPONSE.models[0].name);
 		await page.getByText('Run').click();
 		await expect(
 			page.locator('ol[data-sonner-toaster] li', { hasText: "Can't connect to Ollama server" })
