@@ -6,14 +6,14 @@ import {
 	MOCK_OPENAI_MODELS,
 	MOCK_SESSION_1_RESPONSE_1,
 	mockCompletionResponse,
-	mockOpenAIModelsResponse,
-	mockTagsResponse
+	mockOllamaModelsResponse,
+	mockOpenAIModelsResponse
 } from './utils';
 
 test.describe('FieldSelect', () => {
 	test('filters options, shows selected value, and allows clearing', async ({ page }) => {
 		await page.goto('/settings');
-		await mockTagsResponse(page);
+		await mockOllamaModelsResponse(page);
 		await page.goto('/sessions/qbhc0q');
 
 		const modelCombobox = page.getByLabel('Available models');
@@ -86,7 +86,7 @@ test.describe('FieldSelect', () => {
 	test('models are correctly grouped and sorted', async ({ page }) => {
 		const newSessionButton = page.getByTestId('new-session');
 		await page.goto('/settings');
-		await mockTagsResponse(page);
+		await mockOllamaModelsResponse(page);
 
 		await page.getByText('Sessions', { exact: true }).click();
 		await newSessionButton.click();
@@ -135,7 +135,7 @@ test.describe('FieldSelect', () => {
 
 	test('Ollama models have an ollama badge', async ({ page }) => {
 		await page.goto('/settings');
-		await mockTagsResponse(page);
+		await mockOllamaModelsResponse(page);
 
 		await page.goto('/sessions/new');
 		await page.getByLabel('Available models').click();
